Implment UX for payroll

+++ ios/fyp/TANetworking/src/data/PayslipDetailData.swift
@@ -20,7 +20,7 @@ public struct PayslipDetailData: Codable {
     public let salary: Int?
     public let tax: Double?
     public let ssb: Int?
-    public let overtime: Int?
+    public let overtime: Double?
     public let createdAt: String
     public let deletedAt: String?
     public let acknowledgedAt: String?
